requirements.txt
Flask
pandas
openpyxl
pywin32
pywebview[qt]